Cucumber all in one scenario (#44)
* changed the tests in PointScales to work with an not empty database

* fixed the badge not beeing returned by the api when created or updated. Changed the cucumber badge steps to work with not empty database

* work in progress for the cucumber test of endpoint /rules

* added tests for the rule endpoint

* fixed rule controller not returning the created rule id and fixed the rule cucumber test to check if the api return the id when a rule is created

* fix

* corrected api spec and tests to be able to update an application

* minor fix

* fixed a bug which let all applications access to all rules even if it wasn't there's

* changed the delete method in impl for the rules to return relevant status code. Added a test to check if we can delete a rule

* fixes

+++ impl/src/main/java/ch/heigvd/amt/gamificator/api/rule/RuleService.java
@@ -2,6 +2,7 @@ package ch.heigvd.amt.gamificator.api.rule;
 
 import ch.heigvd.amt.gamificator.api.model.*;
 import ch.heigvd.amt.gamificator.entities.*;
+import ch.heigvd.amt.gamificator.exceptions.NotAuthorizedException;
 import ch.heigvd.amt.gamificator.exceptions.NotFoundException;
 import ch.heigvd.amt.gamificator.exceptions.RelatedObjectNotFound;
 import ch.heigvd.amt.gamificator.repositories.*;
@@ -9,6 +10,7 @@ import lombok.AllArgsConstructor;
 import lombok.Synchronized;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.HttpClientErrorException;
 
 import java.util.LinkedList;
 import java.util.List;
@@ -65,14 +67,26 @@ public class RuleService {
     }
 
 
-    public void delete(Long id) {
-        ruleRepository.deleteById(id);
+    public void delete(Long id, Long applicationId) throws NotFoundException, NotAuthorizedException {
+        List<Rule> rules = ruleRepository.findAllByApplicationId(applicationId);
+        if(rules.size() > 0) {
+            for (Rule rule : rules) {
+                if(rule.getId() == id){
+                    if (rule.getApplication().getId() == applicationId) {
+                        ruleRepository.deleteById(id);
+                        return;
+                    } else {
+                        throw new NotAuthorizedException("This application isn't the owner of the rule");
+                    }
+                }
+            }
+        }
+        throw new NotFoundException("Rule doesn't exist");
     }
 
-    public List<RuleDTO> getAllRules() {
-
+    public List<RuleDTO> getAllRules(Long applicationId) {
         List<RuleDTO> rules = new LinkedList<>();
-        for (Rule rule : ruleRepository.findAll()){
+        for (Rule rule : ruleRepository.findAllByApplicationId(applicationId)){
             RuleDTO ruleDTO = RuleMapper.toDTO(rule);
 
             List<PointsReward> pointsRewards = pointsRewardRepository.findAllByRuleId(rule.getId());
